+++ components/General/Input/SmallInput/index.tsx
@@ -1,5 +1,5 @@
 import React, { ChangeEvent } from 'react';
-import { Children } from 'types';
+import { Children } from 'libs/types';
 import styled from 'styled-components';
 import { NumberInput } from '@components/General';
 import TooltipSelector, { TooltipSelectorProps } from '@components/Tooltips/TooltipSelector';
