+++ components/Trade/Advanced/TradingPanel/Account/AccountModal.tsx
@@ -1,7 +1,7 @@
 import React, { useContext, useCallback, useReducer, useMemo } from 'react';
 import styled from 'styled-components';
 import { NumberSelect, Section } from '@components/General';
-import { UserBalance } from 'types';
+import { UserBalance } from 'libs/types';
 import ErrorComponent from '@components/General/Error';
 import TracerModal from '@components/General/TracerModal';
 import { SlideSelect } from '@components/Buttons';
