+++ components/Trade/Basic/index.tsx
@@ -9,7 +9,7 @@ import styled from 'styled-components';
 import { calcLiquidationPrice, calcNotionalValue } from '@tracer-protocol/tracer-utils';
 import { toApproxCurrency } from '@libs/utils';
 import { Section } from '@components/General';
-import { UserBalance } from 'types';
+import { UserBalance } from 'libs/types';
 import Error from '@components/General/Error';
 import { BigNumber } from 'bignumber.js';
 import { defaults } from '@libs/Tracer';
