+++ context/Web3Context.tsx
@@ -2,7 +2,7 @@ import React, { useEffect, useState } from 'react';
 import Web3 from 'web3';
 import Web3Modal from 'web3modal';
 import WalletConnectProvider from '@walletconnect/web3-provider';
-import { Children } from 'types';
+import { Children } from 'libs/types';
 import { useMachine } from '@xstate/react';
 
 import { web3Machine } from '@machines/Web3Machine';
