+++ libs/Ome/index.ts
@@ -1,5 +1,5 @@
 import { OMEOrder } from '@tracer-protocol/tracer-utils';
-import { Result } from 'types/General';
+import { Result } from 'libs/types/General';
 
 const BASE_URL = process.env.NEXT_PUBLIC_OME_BASE_URL || 'http://localhost:8989';
