+++ libs/types/General.ts
